+++ vue/src/components/Dialogs/Setup/DeploymentPackages/UpdateDialogs/DeploymentPackageUpdateEnvironmentVariableDialog.vue
@@ -38,6 +38,10 @@ const variables = ref<Variable[]>([
         code: "${database.host}"
     },
     {
+        name: "Database Port",
+        code: "${database.port}"
+    },
+    {
         name: "Database Name",
         code: "${database.name}"
     },