+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Enhancements
 * Updated `4spacesdk/ci4authextension` from v1.2.1 to v1.2.2
+* Add `database.port` as env var variable [Issue #59](https://github.com/4spacesdk/kubernetes-service-orchestrator/issues/59)
 
 ### Upgrade guide
 1. Deploy new image